Refactor and docs
